MTP-4377: Refactor to support file uploading progress
