+++ .gitignore
@@ -2,3 +2,4 @@ node_modules/
 __coverage-*__/
 npm-debug.log
 .idea/
+yarn.lock
