+++ lib/cookieJar.js
