+++ lib/globalFetch.js
