+++ lib/progressPlugBrowser.js
