+++ lib/progressPlugNode.js
