+++ lib/uriParser.js
