+++ progressPlugFactory.js
@@ -1,5 +1,5 @@
-import { ProgressPlugBrowser } from './progressPlugBrowser.js';
-import { ProgressPlugNode } from './progressPlugNode.js';
+import { ProgressPlugBrowser } from './lib/progressPlugBrowser.js';
+import { ProgressPlugNode } from './lib/progressPlugNode.js';
 export const progressPlugFactory = {
     get ProgressPlug() {
         if(typeof window === 'undefined') {
